cosmetic bitrate rounding
git-svn-id: http://svn.code.sf.net/p/megui/code@1897 a21b8521-4df1-4ed3-b48c-cff5bc73dcb3

+++ megui/trunk/packages/tools/autoencode/AutoEncodeWindow.cs
@@ -568,7 +568,7 @@ namespace MeGUI
                 data.CalcByTotalSize();
 
                 this.videoSize.Text = data.VideoSize.ToString();
-                this.projectedBitrateKBits.Text = data.VideoBitrate.ToString();
+                this.projectedBitrateKBits.Text = ((int)data.VideoBitrate).ToString();
             }
             catch (Exception)
             {
